moved the randomHexGenerator function to utilities folder
